refactor(app): extract question keyboard-nav from dock
Move the focus/keyboard cluster (clamp, pickFocus, focus, move, nav) out
of session-question-dock.tsx into createQuestionKeyboardNav(deps). The
factory owns focusFrame and its rAF cleanup; the dock injects store,
accessors, a resolveFocusTarget callback over its refs, and next/reject
thunks. focus is now owned by the nav module and injected into the Q4
answer-editing factory's focus dependency.
